+++ FastAnimations/release-notes.md
@@ -2,11 +2,9 @@
 
 # Release notes
 ## 1.7
-* Added support for the load game menu delay.
-
-## 1.6
 * Updated for Stardew Valley 1.3 (including multiplayer support).
 * Added support for title menu transitions.
+* Added support for the blinking-slot delay after you click a load-save slot.
 
 ## 1.5
 * Updated to SMAPI 2.4.